Renamed things from Basket to Zap for better referencing.
