--- src/components/Baskets/index.jsx
@@ -1,5 +1,0 @@
-// Major Work in Progress.
-// Might not be accomplished until the end of the hackathon. Possible that it shall be delayed until later
-import BasketsContainer from './BasketsContainer';
-
-export default BasketsContainer;
