+++ src/components/Zaps/index.jsx
@@ -1,0 +1,3 @@
+import ZapsContainer from './ZapsContainer';
+
+export default ZapsContainer;
